# app/Http/Controllers/API/V1/SubscriptionsController.php
<?php

declare(strict_types=1);

namespace App\Http\Controllers\API\V1;

use App\Models\Plan;
use App\Models\Price;
use Hypervel\Http\Request;
use App\Models\Subscription;
use OpenApi\Attributes as OAT;
use App\OpenApi\Responses\Http400;
use App\OpenApi\Responses\Http401;
use App\Http\Resources\PlanResource;
use App\Services\SubscriptionService;
use Psr\Http\Message\ResponseInterface;
use App\Exceptions\NotFoundHttpException;
use App\Validators\SubscriptionValidator;
use App\Exceptions\InvalidRequestException;
use App\Services\PaddleSubscriptionService;
use App\Services\StripeSubscriptionService;
use App\Http\Controllers\AbstractController;
use App\OpenApi\Schemas\PlanResource as PlanSchema;
use App\OpenApi\Parameters\Path\SubscriptionId as SubscriptionIdParam;

class SubscriptionsController extends AbstractController
{
    #[OAT\Get(
        path: '/v1/subscriptions',
        operationId: 'api.v1.subscriptions.index',
        summary: "Get user's current subscription plan",
        security: [['bearerAuth' => []]],
        tags: ['Subscriptions'],
        responses: [
            new OAT\Response(
                response: 200,
                description: 'Current subscription plan with status',
                content: new OAT\JsonContent(
                    allOf: [
                        new OAT\Schema(
                            properties: [
                                new OAT\Property(
                                    property: 'status',
                                    type: 'string',
                                    nullable: true,
                                    enum: ['trial', 'active'],
                                    example: 'trial',
                                    description: 'trial = on trial; active = paid; null = free plan'
                                ),
                                new OAT\Property(
                                    property: 'trial_ends_at',
                                    type: 'string',
                                    format: 'date-time',
                                    nullable: true,
                                    example: '2026-06-14T00:00:00+00:00',
                                    description: 'Present only when status=trial'
                                ),
                            ]
                        ),
                        new OAT\Schema(ref: PlanSchema::class),
                    ]
                )
            ),
            new OAT\Response(ref: Http401::class, response: 401),
        ]
    )]
    public function index(Request $request, SubscriptionService $subscriptionService): ResponseInterface
    {
        $subscription = $subscriptionService->getUserSubscription($request->user()->id);
        $plan = $subscriptionService->getUserSubscriptionPlan($subscription);

        $plan->load([
            'prices' => function ($builder) use ($subscription) {
                $subscription
                    ? $builder->where('id', $subscription->price_id)
                    : $builder->where('unit', Price::UNIT_MONTHLY)->where('price', 0);
            },
        ]);

        $status = $subscription?->status;
        $trialEndsAt = ($status === Subscription::STATUS_TRIAL)
            ? $subscription->next_date?->toIso8601String()
            : null;

        return response()->json([
            'status'        => $status,
            'trial_ends_at' => $trialEndsAt,
            ...(new PlanResource($plan))->toArray(),
        ]);
    }

    /**
     * @throws InvalidRequestException
     */
    #[OAT\Post(
        path: '/v1/subscriptions',
        operationId: 'api.v1.subscriptions.store',
        summary: 'Initiate a checkout for a subscription',
        security: [['bearerAuth' => []]],
        requestBody: new OAT\RequestBody(
            required: true,
            content: new OAT\JsonContent(
                required: ['planId', 'priceId'],
                properties: [
                    new OAT\Property(
                        property: 'planId',
                        description: 'Plan ID (ULID)',
                        type: 'string',
                        example: '01JCXYZ123456789ABCDEFGHIJ'
                    ),
                    new OAT\Property(
                        property: 'priceId',
                        description: 'Price ID (ULID)',
                        type: 'string',
                        example: '01JCXYZ123456789ABCDEFGHIJ'
                    ),
                    new OAT\Property(
                        property: 'paymentMethod',
                        description: 'Payment gateway (stripe or paddle, defaults to stripe)',
                        type: 'string',
                        enum: ['stripe', 'paddle'],
                        example: 'stripe'
                    ),
                ]
            )
        ),
        tags: ['Subscriptions'],
        responses: [
            new OAT\Response(
                response: 200,
                description: 'Checkout initialization payload',
                content: new OAT\JsonContent(
                    properties: [
                        new OAT\Property(
                            property: 'stripe',
                            properties: [
                                new OAT\Property(property: 'publishable_key', type: 'string', example: 'pk_live_...'),
                                new OAT\Property(
                                    property: 'client_secret',
                                    description: 'Checkout Session client_secret for stripe.initEmbeddedCheckout()',
                                    type: 'string',
                                    example: 'cs_live_xxx_secret_xxx'
                                ),
                            ],
                            type: 'object'
                        ),
                    ]
                )
            ),
            new OAT\Response(ref: Http400::class, response: 400),
            new OAT\Response(ref: Http401::class, response: 401),
        ]
    )]
    public function store(Request $request): ResponseInterface
    {
        $params = $request->only(['planId', 'priceId', 'paymentMethod']);

        $v = new SubscriptionValidator($params);
        $v->setStoreRules();

        if (!$v->passes()) {
            throw new InvalidRequestException($v->errors()->toArray());
        }

        if (!$plan = Plan::query()->find($params['planId'])) {
            throw new InvalidRequestException(['planId' => [__('validators.controllers.subscription.plan_not_found')]]);
        }

        if (!$price = Price::query()->find($params['priceId'])) {
            throw new InvalidRequestException(
                ['priceId' => [__('validators.controllers.subscription.price_not_found')]]
            );
        }

        if (!$plan->prices()->find($price->id)) {
            throw new InvalidRequestException(
                ['priceId' => [__('validators.controllers.subscription.price_not_in_plan')]]
            );
        }

        $paymentMethod = $params['paymentMethod'] ?? Subscription::PAYMENT_METHOD_STRIPE;

        $subscription = $request->user()->subscriptions()->create([
            'plan_id'        => $plan->id,
            'price_id'       => $price->id,
            'payment_method' => $paymentMethod,
            'status'         => Subscription::STATUS_PAYING,
        ]);

        if ($paymentMethod === Subscription::PAYMENT_METHOD_STRIPE) {
            $data = (new StripeSubscriptionService())->createCheckout(
                $request->user(),
                $plan,
                $price,
                $subscription
            );
        } else {
            $data = (new PaddleSubscriptionService())->createCheckout(
                $request->user(),
                $plan,
                $price,
                $subscription
            );
        }

        return response()->json($data);
    }

    /**
     * @throws InvalidRequestException
     */
    #[OAT\Put(
        path: '/v1/subscriptions/{subscriptionId}',
        operationId: 'api.v1.subscriptions.update',
        summary: 'Confirm subscription after successful Paddle payment',
        security: [['bearerAuth' => []]],
        requestBody: new OAT\RequestBody(
            required: true,
            content: new OAT\JsonContent(
                required: ['transaction_id'],
                properties: [
                    new OAT\Property(
                        property: 'transaction_id',
                        description: 'Paddle transaction ID',
                        type: 'string',
                        example: 'txn_01abc...'
                    ),
                ]
            )
        ),
        tags: ['Subscriptions'],
        parameters: [
            new OAT\Parameter(ref: SubscriptionIdParam::class),
        ],
        responses: [
            new OAT\Response(response: 200, description: 'OK'),
            new OAT\Response(ref: Http400::class, response: 400),
            new OAT\Response(ref: Http401::class, response: 401),
        ]
    )]
    public function update(Request $request, string $subscriptionId)
    {
        if (!$subscription = $request->user()->subscriptions()->find($subscriptionId)) {
            throw new InvalidRequestException(
                ['subscriptionId' => [__('validators.controllers.subscription.not_found')]]
            );
        }

        $confirmed = (new PaddleSubscriptionService())->confirm(
            $subscription,
            $request->input('transaction_id', '')
        );

        if ($confirmed) {
            return response()->make(self::RESPONSE_OK);
        }
    }

    /**
     * 取消訂閱.
     */
    #[OAT\Delete(
        path: '/v1/subscriptions/{subscriptionId}',
        operationId: 'api.v1.subscriptions.destroy',
        summary: 'Cancel subscription',
        security: [['bearerAuth' => []]],
        tags: ['Subscriptions'],
        parameters: [
            new OAT\Parameter(ref: SubscriptionIdParam::class),
        ],
        responses: [
            new OAT\Response(response: 200, description: 'OK'),
            new OAT\Response(ref: Http401::class, response: 401),
            new OAT\Response(response: 404, description: 'Subscription not found'),
        ]
    )]
    public function destroy(Request $request, SubscriptionService $subscriptionService)
    {
        if (!$subscription = $subscriptionService->getUserSubscription($request->user()->id)) {
            throw new NotFoundHttpException();
        }

        if ($subscription->payment_method === Subscription::PAYMENT_METHOD_STRIPE) {
            (new StripeSubscriptionService())->cancel($subscription);
        } else {
            (new PaddleSubscriptionService())->cancel($subscription);
        }

        return response()->make(self::RESPONSE_OK);
    }
}